requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1

# LangChain / LangGraph
langchain==0.3.7
langchain-openai==0.2.9
langchain-anthropic==0.3.0
langgraph==0.2.53

# Auth
python-jose[cryptography]==3.3.0
